rss-reader/requirements.txt
alembic==1.7.1
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
factory-boy==3.2.1
Faker==9.8.0
fastapi==0.68.1
feedparser==6.0.8
greenlet==1.1.1
h11==0.12.0
idna==3.3
iniconfig==1.1.1
kombu==5.1.0
lxml==4.6.4
Mako==1.1.5
MarkupSafe==2.0.1
packaging==21.0
pluggy==1.0.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
py==1.10.0
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
redis==3.5.3
requests==2.26.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3
SQLAlchemy==1.4.23
starlette==0.14.2
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
vine==5.0.0
wcwidth==0.2.5
